requirements.txt
--constraint https://raw.githubusercontent.com/apache/airflow/constraints-3.1.0/constraints-3.12.txt
minio==7.2.0
tensorflow
torch
configparser==6.0.0
apache-airflow-providers-apache-spark
apache-airflow-providers-standard
scikit-learn
joblib
duckdb
xgboost
deltalake